refactor: give `scheduleProps` its own module
